+++ projectroles/templatetags/projectroles_tags.py
@@ -219,8 +219,10 @@ def is_inherited_owner(project, user):
 
 @register.simple_tag
 def get_app_link_state(app_plugin, app_name, url_name):
-    """Return "active" if plugin matches app_name and url_name is found in
-    app_plugin.urls."""
+    """
+    Return "active" if plugin matches app_name and url_name is found in
+    app_plugin.urls.
+    """
     if app_name == app_plugin.name and url_name in [
         u.name for u in app_plugin.urls
     ]:
@@ -229,20 +231,22 @@ def get_app_link_state(app_plugin, app_name, url_name):
 
 
 @register.simple_tag
-def get_pr_link_state(app_urls, url_name, link_names=None):
-    """Version of get_app_link_state() to be used within the projectroles app.
+def get_pr_link_state(app_urls, request, link_names=None):
+    """
+    Version of get_app_link_state() to be used within the projectroles app.
     If link_names is set, only return "active" if url_name is found in
-    link_names."""
+    link_names.
+    """
+    if request.resolver_match.app_name != 'projectroles':
+        return ''
+    url_name = request.resolver_match.url_name
     if url_name in [u.name for u in app_urls]:
         if link_names:
             if not isinstance(link_names, list):
                 link_names = [link_names]
-
             if url_name not in link_names:
                 return ''
-
         return 'active'
-
     return ''